double -> float as originally was in parse_stl lib. Also added exception handling and some stop-flags.

+++ Graph3D.cpp
@@ -1,6 +1,6 @@
 #include "Graph3D.hpp"
 
-const double Graph3D::INFTY = 20E10;
+const float Graph3D::INFTY = 20E10;
 
 void Graph3D::PlaceEdge(const Point& p1, const Point& p2) {
 	auto search = find(p1);
